+++ pip-services4-components-node/src/run/index.ts
@@ -9,9 +9,6 @@
  */
 export { Parameters } from './Parameters';
 
-export { IContext } from './IContext';
-export { Context } from './Context';
-
 export { IParameterized } from './IParameterized';
 export { IOpenable } from './IOpenable';
 export { IClosable } from './IClosable';